+++ src/main/java/com/voltaire/restaurant/RestaurantService.java
@@ -1,11 +1,13 @@
 package com.voltaire.restaurant;
 
+import com.voltaire.error.customerrors.EntityNotFoundException;
 import com.voltaire.restaurant.model.Restaurant;
 import com.voltaire.restaurant.repository.RestaurantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -22,7 +24,12 @@ public class RestaurantService {
     }
 
     public Restaurant findById(Long id) {
-        return restaurantRepository.findById(id).orElse(null);
+        Optional<Restaurant> restaurant = restaurantRepository.findById(id);
+        if(restaurant.isEmpty()) {
+            throw new EntityNotFoundException(Restaurant.class, "id", id.toString());
+        }
+
+        return restaurant.get();
     }
 
     public boolean notExists(Long id) {
@@ -30,11 +37,19 @@ public class RestaurantService {
     }
 
     public void updateRestaurant(Long id, Restaurant restaurant) {
+        if(notExists(id)) {
+            throw new EntityNotFoundException(Restaurant.class, "id", id.toString());
+        }
+
         restaurant.setId(id);
         restaurantRepository.save(restaurant);
     }
 
     public void deleteRestaurant(Long id) {
+        if(notExists(id)) {
+            throw new EntityNotFoundException(Restaurant.class, "id", id.toString());
+        }
+
         restaurantRepository.deleteById(id);
     }
